requirements.txt
# diffusers[torch]==0.26.3
# diffusers[torch]==0.27.0
git+https://github.com/yhZhai/diffusers.git
transformers==4.36.2
wandb
matplotlib
torchmetrics==1.3.1
torch-fidelity==0.3.0
einops
azure-storage-blob==12.12.0
tensorboard
tensorboardX
ffmpeg-python
opencv-python
timm
ftfy
rouge_score
omegaconf
decord
colorlog
deepdish
configobj
json_lines
albumentations
pudb
imageio
imageio-ffmpeg
pytorch-lightning
omegaconf
test-tube
streamlit
setuptools
kornia
clean-fid
pytorch-fid
h5py
lpips
tabulate
ninja
matplotlib
webdataset
braceexpand
Pillow
accelerate==0.29.3
compel==0.1.8
datasets
filelock
flax>=0.4.1
hf-doc-builder>=0.3.0
huggingface-hub>=0.20.2
requests-mock==1.10.0
importlib_metadata
invisible-watermark>=0.2.0
isort>=5.5.4
jax>=0.4.1
jaxlib>=0.4.1
Jinja2
k-diffusion>=0.0.12
torchsde
note_seq
librosa
numpy
parameterized
git+https://github.com/yhZhai/peft.git
protobuf==3.20.3
pytest
pytest-timeout
pytest-xdist
ruff==0.1.5
safetensors>=0.3.1
sentencepiece>=0.1.91,!=0.1.92
GitPython<3.1.19
# scipy==1.11.1
onnx
regex!=2019.12.17
requests
bitsandbytes
git+https://github.com/microsoft/azfuse.git
deepspeed==0.11.2
# deepspeed==0.6.6
albumentations
mlflow
moviepy
git+https://github.com/openai/CLIP.git
av
git+https://github.com/yhZhai/open_clip.git
